// play/chess-frontend/src/api/def/model/chessMoveWrapperDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Aide axum Open API
 * # Todo API

A very simple Todo server with documentation.

The purpose is to showcase the documentation workflow of Aide rather
than a correct implementation.

 */
import type { ChessMoveDto } from "./chessMoveDto";

export interface ChessMoveWrapperDto {
  confidence: number;
  inner: ChessMoveDto;
  prior: number;
}